mm: Tune dirty data writebacks To relieve memory pressure on Android, we should start writing dirty data sooner than later.
The Linux kernel's default is way too conservative on Android.
Android doesn't have a good reason to delay writing dirty data
anyways.

This change will make the kernel start writing dirty data when
it reaches 50M and switch to synchronous on 200M, which is far
more sensible values than 1.2G and 2.4G respectively on a device
with 12G of RAM.

Change-Id: Ifcb8527edd2c71461b5a29074398a97dd1a158f7
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
[EmanuelCN: Adapt the change for kernel]
Signed-off-by: skye-tachyon <sonicgames772@gmail.com>